+++ app/controllers/search_history_controller.rb
@@ -3,4 +3,6 @@ class SearchHistoryController < ApplicationController #:nodoc:
   include Blacklight::SearchHistory
 
   helper BlacklightAdvancedSearch::RenderConstraintsOverride
+  helper BlacklightRangeLimit::ViewHelperOverride
+  helper RangeLimitHelper
 end