+++ app/models/search_builder.rb
@@ -3,6 +3,8 @@
 # :nodoc:
 class SearchBuilder < Blacklight::SearchBuilder
   include Blacklight::Solr::SearchBuilderBehavior
+  include BlacklightRangeLimit::RangeLimitBuilder
+
   include BlacklightAdvancedSearch::AdvancedSearchBuilder
   self.default_processor_chain += [:add_advanced_parse_q_to_solr, :add_advanced_search_to_solr]
   include BlacklightHeatmaps::SolrFacetHeatmapBehavior